Validator/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
groq==0.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
numpy==2.2.2
pillow==10.4.0
proglog==0.1.10
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yt-dlp==2025.1.26
